Adds shift module and changes decrypt/encrypt class to encrypter/decrypter class
