+++ lib/shift_module.rb
@@ -1,11 +1,4 @@
-class Encrypt
-  attr_reader :keys, :offset, :encrypter
-
-  def initialize(keys, offset)
-    @keys = keys
-    @offset = offset
-  end
-
+module ShiftModule
   def create_letters
     ("a".."z").to_a
   end
@@ -28,22 +21,5 @@ class Encrypt
         encrypter[num] = Hash[create_letters.zip(create_letters.rotate(new_shift[num]))]
         encrypter
       end
-  end
-
-  def encrypt_message(message, key, date)
-    shift_hash(key, date)
-    num = -1
-
-    message.downcase.chars.map do |letter|
-      num += 1
-      if create_letters.include?(letter) && num <= 3
-        @encrypter[num][letter]
-      elsif create_letters.include?(letter) && num > 3
-        num = 0
-        @encrypter[num][letter]
-      else
-        letter
-      end
-    end.join
   end
 end
